Index / Details
Index page has been set to list the teams as required in the specifications.

Details page now display the club image alongside a link for the official club website.

The latter is being dynamically generated by the controllers with the use of viewmodels.

+++ ClubSparkTest/Models/Team.cs
@@ -1,5 +1,6 @@
 ﻿using System;
 using System.Collections.Generic;
+using System.ComponentModel.DataAnnotations;
 using System.Linq;
 using System.Web;
 
@@ -7,7 +8,7 @@ namespace ClubSparkTest.Models
 {
     public class Team
     {
-        public int ID { get; set; }
+        public double ID { get; set; }
         public string Name { get; set; }
         public string Country { get; set; }
         public bool Eliminated { get; set; }
